+++ app/src/main/java/chimera/epic_rpg_thing/model/Classes/WarriorClass.java
@@ -12,11 +12,11 @@ public class WarriorClass extends CharacterClass {
         this.setName("Warrior");
         this.addClassMod("hpMulti", 10);
         this.addClassMod("manaMulti", 3);
-        this.addWeaknesses(new ElementalEffect(ElementalType.POISON, (short)1));
+        this.addWeaknesses(new ElementalEffect(ElementalType.POISON, (int)1));
         initializeSkills();
     }
 
     private void initializeSkills(){
-        this.classSkills.put("BerserkerRage1", AOEFactory.create(AOEType.AOE_DEBUFF, 0,0,0, null, new ElementalEffect(ElementalType.PHYSICAL, (short)1), "With an insane roar the warrior terrifies there enemies.", 4));
+        this.classSkills.put("BerserkerRage1", AOEFactory.create(AOEType.AOE_DEBUFF, 0,0,0, null, new ElementalEffect(ElementalType.PHYSICAL, (int)1), "With an insane roar the warrior terrifies there enemies.", 4));
     }
 }
